# server/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: felixtj.dietrich@tum.de
    name: Felix T.J. Dietrich
  description: API documentation for the Hephaestus application server.
  license:
    name: MIT License
    url: https://github.com/ls1intum/Hephaestus/blob/develop/LICENSE
  title: Hephaestus API
  version: 0.0.0-development
servers:
- description: Default Server URL
  url: /
security:
- bearerAuth: []
...
paths:
...
  /workspaces/{workspaceSlug}/team:
    get:
      description: Returns all teams in the workspace organization
      operationId: getAllTeams
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamInfo"
          description: list of teams with their members and permissions
      security: []
      summary: List teams
      tags:
      - Teams
  /workspaces/{workspaceSlug}/team/{id}/visibility:
    post:
      description: Show or hide a team in leaderboard calculations
      operationId: updateTeamVisibility
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      - description: "whether to hide the team (query parameter, fallback)"
        in: query
        name: hidden
        required: false
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
        description: "whether to hide the team (body parameter, preferred)"
      responses:
        "200":
          description: "200 OK on success, 404 if team not found"
      summary: Update team visibility
      tags:
      - Teams
  /workspaces/{workspaceSlug}/team/{teamId}/repositories/{repositoryId}/visibility:
    post:
      operationId: updateRepositoryVisibility
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: hiddenFromContributions
        required: false
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
      responses:
        "200":
          description: OK
      tags:
      - Teams
...
components:
  schemas:
...
    LabelInfo:
      type: object
      description: Information about a label from a repository
      properties:
        color:
          type: string
          description: "Hex color code of the label (without #)"
          example: d73a4a
        id:
          type: integer
          format: int64
          description: Unique identifier of the label
        name:
          type: string
          description: Name of the label
        repository:
          $ref: "#/components/schemas/RepositoryInfo"
          description: Repository the label belongs to
      required:
      - color
      - id
      - name
...
    RepositoryInfo:
      type: object
      description: Information about a git repository
      properties:
        description:
          type: string
          description: Description of the repository
        hiddenFromContributions:
          type: boolean
          description: Whether contributions from this repository are hidden from
            leaderboard calculations
        htmlUrl:
          type: string
          description: URL to the repository on the git provider
        id:
          type: integer
          format: int64
          description: Unique identifier of the repository
        labels:
          type: array
          description: Labels defined in the repository
          items:
            $ref: "#/components/schemas/LabelInfo"
        name:
          type: string
          description: Name of the repository
        nameWithOwner:
          type: string
          description: "Full name including owner (e.g., 'owner/repo')"
      required:
      - hiddenFromContributions
      - htmlUrl
      - id
      - name
      - nameWithOwner
...
    TeamInfo:
      type: object
      description: "Detailed information about a team including members, repositories,\
        \ and labels"
      properties:
        description:
          type: string
          description: Description of the team
        hidden:
          type: boolean
          description: Whether the team is hidden from leaderboard display
        htmlUrl:
          type: string
          description: URL to the team's page on the git provider
        id:
          type: integer
          format: int64
          description: Unique identifier of the team
        labels:
          type: array
          description: Labels configured as filters for this team
          items:
            $ref: "#/components/schemas/LabelInfo"
        members:
          type: array
          description: Members of the team (excluding bots)
          items:
            $ref: "#/components/schemas/UserInfo"
        membershipCount:
          type: integer
          format: int32
          description: Total count of team memberships
        name:
          type: string
          description: Name of the team
        organization:
          type: string
          description: Organization the team belongs to
        parentId:
          type: integer
          format: int64
          description: "ID of the parent team, if this is a sub-team"
        privacy:
          type: string
          description: Privacy level of the team (SECRET or VISIBLE)
          enum:
          - SECRET
          - VISIBLE
        repoPermissionCount:
          type: integer
          format: int32
          description: Count of repository permissions
        repositories:
          type: array
          description: Repositories the team has access to
          items:
            $ref: "#/components/schemas/RepositoryInfo"
      required:
      - hidden
      - id
      - labels
      - members
      - membershipCount
      - name
      - repoPermissionCount
      - repositories
...
    UserInfo:
      type: object
      description: Information about a user from the git provider
      properties:
        avatarUrl:
          type: string
          description: URL to the user's avatar image
        email:
          type: string
          description: "Email address of the user, if public"
        htmlUrl:
          type: string
          description: URL to the user's profile on the git provider
        id:
          type: integer
          format: int64
          description: Unique identifier of the user
        leaguePoints:
          type: integer
          format: int32
          description: League points earned by the user in the current scope
          example: 150
        login:
          type: string
          description: Login/username of the user
        name:
          type: string
          description: Display name of the user
      required:
      - avatarUrl
      - htmlUrl
      - id
      - login
      - name
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Hephaestus-native JWT bearer authentication. The SPA normally authenticates
        via the `__Host-HEPHAESTUS_AT` session cookie; this scheme documents the equivalent
        bearer token.
      scheme: bearer
      type: http